pyproject.toml
[project]
name = "repl-nix-studynexus"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "mammoth>=1.8.0",
    "openai>=1.58.1",
    "pdf2image>=1.17.0",
    "pillow>=11.0.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
]
